Fix bugs and attach implementation for pillars, hero's interaction and gameLoop

+++ src/Thief.java
@@ -15,7 +15,7 @@ public class Thief extends Hero{
      * @param theBlock
      * @param theCritChance
      */
-    public Thief(Double theHit, int theAttack, String theName, int theMin, int theMax, int theBlock, int theCritChance) {
+    public Thief(Double theHit, int theAttack, String theName, int theMin, int theMax, Double theBlock, double theCritChance) {
         super(theHit, theAttack, theName, theMin, theMax, theBlock, theCritChance);
     }
